[Chore] Migrate V0 attention utils
